+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * IniFile::content() should not fail if file doesn't exist
     * Session: Protection against invalid session cookie name throwing exception
     * Session: Do not destroy session on CLI
+    * BlueprintSchema: Fixed warning when field list is not what was expected 
 
 # v1.3.3
 ## 10/06/2016